@@ -79,7 +79,7 @@
         "@swc/core": "^1.3.107",
         "@swc/jest": "^0.2.36",
         "@testing-library/jest-dom": "^6.4.2",
-        "@testing-library/react": "^15.0.0",
+        "@testing-library/react": "^15.0.5",
         "@types/chrome": "0.0.260",
         "@types/jest": "^29.5.11",
         "@types/jsdom": "^21.1.6",
@@ -160,8 +160,8 @@
         "idb-keyval": "^6.2.1",
         "lodash": "^4.17.21",
         "luxon": "^3.4.4",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
         "react-helmet": "^6.1.0",
         "react-resize-detector": "^9.1.1",
         "react-router-dom": "^6.21.3",
