@@ -23,8 +23,8 @@
         "classnames": "^2.5.1",
         "lodash": "^4.17.21",
         "luxon": "^3.4.4",
-        "react": "^16.14.0",
-        "react-dom": "^16.14.0",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
         "react-helmet": "^6.1.0",
         "uuid": "^9.0.1"
     }
