@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { Dialog, DialogFooter } from '@fluentui/react';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import { ToolData } from 'common/types/store-data/unified-data-interface';
 import * as React from 'react';
 import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
@@ -133,7 +133,10 @@ describe('IssueFilingDialog', () => {
             ActionAndCancelButtonsComponent,
         );
 
-        actionCancelButtons.cancelButtonOnClick(null);
+        act(() => {
+            actionCancelButtons.cancelButtonOnClick(null);
+        })
+
         isSettingsValidMock.verifyAll();
         onCloseMock.verifyAll();
     });
@@ -162,7 +165,10 @@ describe('IssueFilingDialog', () => {
         const actionCancelButtons = getMockComponentClassPropsForCall(
             ActionAndCancelButtonsComponent,
         );
-        actionCancelButtons.primaryButtonOnClick(eventStub);
+        act(() => {
+            actionCancelButtons.primaryButtonOnClick(eventStub);
+        })
+
 
         isSettingsValidMock.verifyAll();
         userConfigMessageCreatorMock.verifyAll();
@@ -200,7 +206,10 @@ describe('IssueFilingDialog', () => {
             propertyName: propertyStub,
             propertyValue: propertyValueStub,
         };
-        issueFilingSettingsContainer.onPropertyUpdateCallback(payload);
+        act(() => {
+            issueFilingSettingsContainer.onPropertyUpdateCallback(payload);
+        })
+
 
         expect(renderResult.asFragment()).toMatchSnapshot();
         expectMockedComponentPropsToMatchSnapshots([Dialog]);
@@ -228,7 +237,10 @@ describe('IssueFilingDialog', () => {
             propertyName: propertyStub,
             propertyValue: propertyValueStub,
         };
-        issueFilingSettingsContainer.onPropertyUpdateCallback(payload);
+        act(() => {
+            issueFilingSettingsContainer.onPropertyUpdateCallback(payload);
+        })
+
 
         expect(renderResult.asFragment()).toMatchSnapshot();
         expectMockedComponentPropsToMatchSnapshots([Dialog]);
@@ -264,7 +276,10 @@ describe('IssueFilingDialog', () => {
         const payload = {
             issueFilingServiceName: differentServiceKey,
         };
-        issueFilingSettingsContainer.onSelectedServiceChange(payload);
+        act(() => {
+            issueFilingSettingsContainer.onSelectedServiceChange(payload);
+        })
+
 
         expect(renderResult.asFragment()).toMatchSnapshot();
         expectMockedComponentPropsToMatchSnapshots([Dialog]);
