@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import * as React from 'react';
 import { IMock, It, Mock, Times } from 'typemoq';
 
@@ -27,8 +27,10 @@ describe('ToastTest', () => {
         const renderResult = render(<Toast ref={toastRef} {...props}></Toast>);
 
         expect(renderResult.asFragment()).toMatchSnapshot('render nothing before show() is called');
+        act(() => {
+            toastRef.current.show('hello world');
+        })
 
-        toastRef.current.show('hello world');
 
         expect(renderResult.asFragment()).toMatchSnapshot('render content');
     });
