@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import {
     TelemetryViewer,
     TelemetryViewerDeps,
@@ -72,14 +72,16 @@ describe('TelemetryViewer', () => {
                 .callback(l => (callback = l));
 
             const renderResult = render(<TelemetryViewer {...props} />);
+            act(() => {
+                callback({
+                    key: 'value',
+                    applicationBuild: 'test-application-build',
+                    applicationName: 'test-application-name',
+                    applicationVersion: 'test-application-version',
+                    installationId: 'test-installation-id',
+                });
+            })
 
-            callback({
-                key: 'value',
-                applicationBuild: 'test-application-build',
-                applicationName: 'test-application-name',
-                applicationVersion: 'test-application-version',
-                installationId: 'test-installation-id',
-            });
 
             expect(renderResult.asFragment()).toMatchSnapshot();
             expectMockedComponentPropsToMatchSnapshots([TelemetryMessagesList]);
