@@ -1,15 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { Theme } from 'common/components/theme';
 import { configMutator } from 'common/configuration';
-import * as React from 'react';
-import { InsightsRouter } from 'views/insights/insights-router';
+import { createRoot } from 'react-dom/client';
 import { renderer, RendererDeps } from 'views/insights/renderer';
 
+jest.mock('../../../../../common/components/body-class-modifier');
+jest.mock('common/components/theme');
 describe('insights renderer', () => {
     const deps = {
         dom: document,
-        render: jest.fn(),
+        render: jest.fn(createRoot),
         initializeFabricIcons: jest.fn(),
     } as Partial<RendererDeps> as RendererDeps;
 
@@ -34,12 +34,6 @@ describe('insights renderer', () => {
     it('renders InsightsRouter', () => {
         renderer(deps);
         const root = document.body.querySelector('#insights-root');
-        expect(deps.render).toHaveBeenCalledWith(
-            <>
-                <Theme deps={deps} />
-                <InsightsRouter deps={deps} />
-            </>,
-            root,
-        );
+        expect(deps.render).toHaveBeenCalledWith(root);
     });
 });
